python-worker/requirements.txt
# Core dependencies
redis==5.0.1
rq==1.15.1
python-dotenv==1.0.0

# Text processing
regex==2023.10.3
langdetect==1.0.9
ftfy==6.1.3
git+https://github.com/haven-jeon/PyKoSpacing.git
tensorflow==2.20.0
numpy>=2.3.0
kss==6.0.5

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# gRPC (optional, for faster cache service)
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf>=5.29.0  # gRPC 호환 버전

# Logging
loguru==0.7.2

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Performance
orjson==3.9.10
